wip move lang buttons and fix
