+++ src/App.tsx
@@ -9,10 +9,12 @@ import {
 } from './container';
 import { Navbar } from './components';
 import './App.scss';
+import Languages from './components/Languages/Languages.tsx';
 
 const App = () => {
   return (
     <div className="app">
+      <Languages />
       <Navbar />
       <Header />
       <About />
